+++ src/main/scala/com/rayyildiz/sentiment_analyzer/actors/NLPActor.scala
@@ -4,12 +4,12 @@ import akka.actor.SupervisorStrategy.{Escalate, Restart, Resume}
 import akka.actor.{Actor, OneForOneStrategy, Props, SupervisorStrategy}
 import com.google.cloud.language.v1.LanguageServiceClient
 import com.google.cloud.translate.Translate
-import com.rayyildiz.sentiment_analyzer.models._
 
 import scala.concurrent.TimeoutException
 import scala.concurrent.duration._
 
 class NLPActor(languageApi: LanguageServiceClient, translateApi: Translate) extends Actor {
+  import com.rayyildiz.sentiment_analyzer.models.TextModels._
 
   override def receive: Receive = {
     case ExtractWord(text) =>
