+++ src/main/scala/com/rayyildiz/sentiment_analyzer/controllers/ApplicationController.scala
@@ -1,10 +1,11 @@
 package com.rayyildiz.sentiment_analyzer.controllers
 
 import javax.inject.Singleton
-import com.rayyildiz.sentiment_analyzer.models.{ApplicationInformation, ApplicationStatus, Ping, Pong}
+import com.rayyildiz.sentiment_analyzer.models.ApplicationModels.{ApplicationInformation, ApplicationStatus, Ping, Pong}
 
 @Singleton
 class ApplicationController {
+
   def ping(ping: Ping): Pong = Pong(ping.message)
 
   def health: ApplicationStatus = ApplicationStatus(status = true, name = "sentiment_analyzer")
