+++ src/main/scala/com/rayyildiz/sentiment_analyzer/models/ApplicationModels.scala
@@ -1,12 +1,16 @@
 package com.rayyildiz.sentiment_analyzer.models
 
-case class Ping(message: String)
+object ApplicationModels {
 
-case class Pong(message: String)
+  case class Ping(message: String)
 
-// Error Message
-case class ErrorMessage(statusCode: Int, error: String)
+  case class Pong(message: String)
 
-case class ApplicationStatus(status: Boolean, name: String)
+  // Error Message
+  case class ErrorMessage(statusCode: Int, error: String)
 
-case class ApplicationInformation(version: String, name: String)
+  case class ApplicationStatus(status: Boolean, name: String)
+
+  case class ApplicationInformation(version: String, name: String)
+
+}
