+++ src/main/scala/com/rayyildiz/sentiment_analyzer/models/TextModels.scala
@@ -1,24 +1,28 @@
 package com.rayyildiz.sentiment_analyzer.models
 
-// Clean text
-case class CleanText(text: String)
+object TextModels {
 
-// Extract Important Words
-case class ExtractWord(text: String)
+  // Clean text
+  case class CleanText(text: String)
 
-// Extract Words Indicating Sentiment
-case class SentimentWord(text: String)
+  // Extract Important Words
+  case class ExtractWord(text: String)
 
-// Language detect text
-case class LanguageDetectText(text: String)
+  // Extract Words Indicating Sentiment
+  case class SentimentWord(text: String)
 
-case class DeterminationRelationWord(text: String)
+  // Language detect text
+  case class LanguageDetectText(text: String)
 
-case class DeterminationToken(
-    lemma: String,
-    tag: String,
-    number: Option[String],
-    person: Option[String],
-    mood: Option[String],
-    tense: Option[String]
-)
+  case class DeterminationRelationWord(text: String)
+
+  case class DeterminationToken(
+      lemma: String,
+      tag: String,
+      number: Option[String],
+      person: Option[String],
+      mood: Option[String],
+      tense: Option[String]
+  )
+
+}
